Refactoring tests for github_agent
