+++ tests/unit/github_agent/test_create_checkout_branch.py
@@ -1,0 +1,24 @@
+from unittest.mock import patch, call, MagicMock
+
+@patch("osa_tool.github_agent.github_agent.logger")
+def test_create_and_checkout_branch_exists(mock_logger, github_agent):
+    # Arrange
+    github_agent.repo.heads = {"feature-branch": MagicMock()}
+    # Act
+    github_agent.create_and_checkout_branch()
+    # Assert
+    github_agent.repo.git.checkout.assert_called_once_with("feature-branch")
+    mock_logger.info.assert_called_once_with("Branch feature-branch already exists. Switching to it...")
+
+@patch("osa_tool.github_agent.github_agent.logger")
+def test_create_and_checkout_branch_new(mock_logger, github_agent):
+    # Arrange
+    github_agent.repo.heads = {}
+    # Act
+    github_agent.create_and_checkout_branch()
+    # Assert
+    github_agent.repo.git.checkout.assert_called_once_with('-b', github_agent.branch_name)
+    mock_logger.info.assert_has_calls([
+        call(f"Creating and switching to branch {github_agent.branch_name}..."),
+        call(f"Switched to branch {github_agent.branch_name}.")
+    ])
